agg seccion MiInformacion, optim. validaciones

+++ src/main/java/UI/Tools.java
@@ -96,7 +96,7 @@ public class Tools {
 
         DateTimeFormatter formato = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm");
         actualizarTabla(tabla, transaccionesList,
-                t -> new Object[]{t.getFecha().format(formato), t.getTipo(), (double) t.getMonto(), t.getDescripcion(), t.getNumCuenta()},
+                t -> new Object[]{t.getFecha().format(formato), t.getTipo(), (double) t.getMonto(), t.getDescripcion(), (int)t.getNumCuenta()},
                 new Object[]{"No hay transacciones"});
     }
